Update v0.3.0
上位机：
1. 重构了显示日志逻辑，增加关键日志信息过滤，解决串口通信频率过高导致的日志难以读取问题。
2. 新增清扫图标，更换了数据库相关操作的图标。
3. 实现数据库的关闭，另存为功能。
4. 串口线程新增一个定时器，若无通信10s则关闭信道（不是串口，是通信协议里的信道），且可以修正主机异常重启导致的日志颜色问题。
5. 从机类型名字命名添加下划线支持。
6. 新增启动时窗口居中显示。
host（主机）- 12c5a60s2：
1. 优化了通信协议。
2. 修复越界导致延时的bug。
3. 成功完成 PC->HOST->SLAVE 再 SLAVE->HOST->PC 的通信过程。
skave（从机）：
3. 优化了通信协议，同时支持10位异步通信和11位异步通信。

+++ proj/src/serialport_2.c
@@ -54,7 +54,7 @@ void SP_QTransmitData(uchar* buf, uchar n, bit second)
 #endif
 
 #ifdef USE_QUERY_RECEIVE
-bit SP1_QRWait(uint t) // 11.0592 约1ms
+bit SP1_QRWait(uchar t) // 11.0592 约1ms
 {
     unsigned char i, j;
     if (RI)
